Fix Enable All breakpoint option and re factor All enablement code
This commit refactors existing Enable All & Disable All code to remove
redundant statements.
