+++ debug/org.eclipse.debug.ui/ui/org/eclipse/debug/internal/ui/actions/breakpoints/DisableAllBreakpointsAction.java
@@ -15,30 +15,21 @@ package org.eclipse.debug.internal.ui.actions.breakpoints;
 
 import org.eclipse.core.resources.IMarkerDelta;
 import org.eclipse.core.runtime.CoreException;
-import org.eclipse.core.runtime.IProgressMonitor;
-import org.eclipse.core.runtime.IStatus;
-import org.eclipse.core.runtime.Status;
-import org.eclipse.core.runtime.jobs.Job;
 import org.eclipse.debug.core.DebugPlugin;
-import org.eclipse.debug.core.IBreakpointManager;
 import org.eclipse.debug.core.IBreakpointsListener;
 import org.eclipse.debug.core.model.IBreakpoint;
 import org.eclipse.debug.internal.ui.DebugUIPlugin;
-import org.eclipse.debug.internal.ui.actions.AbstractRemoveAllActionDelegate;
-import org.eclipse.debug.internal.ui.actions.ActionMessages;
-import org.eclipse.debug.internal.ui.preferences.IDebugPreferenceConstants;
+import org.eclipse.debug.internal.ui.actions.AbstractAllBreakpointEnablement;
 import org.eclipse.jface.action.IAction;
-import org.eclipse.jface.dialogs.IDialogConstants;
-import org.eclipse.jface.dialogs.MessageDialogWithToggle;
-import org.eclipse.jface.preference.IPreferenceStore;
-import org.eclipse.ui.IWorkbenchWindow;
 
-public class DisableAllBreakpointsAction extends AbstractRemoveAllActionDelegate implements IBreakpointsListener {
+public class DisableAllBreakpointsAction extends AbstractAllBreakpointEnablement implements IBreakpointsListener {
 
 	@Override
 	protected boolean isEnabled() {
-		IBreakpointManager breakpointManager = DebugPlugin.getDefault().getBreakpointManager();
-		IBreakpoint[] breakpoints = breakpointManager.getBreakpoints();
+		IBreakpoint[] breakpoints = getBreakpoints();
+		if (breakpoints.length == 1) {
+			return false;
+		}
 		for (IBreakpoint bp : breakpoints) {
 			try {
 				if (bp.isEnabled()) {
@@ -50,7 +41,6 @@ public class DisableAllBreakpointsAction extends AbstractRemoveAllActionDelegate
 		}
 		return false;
 	}
-
 	@Override
 	public void breakpointsAdded(IBreakpoint[] breakpoints) {
 		update();
@@ -82,44 +72,6 @@ public class DisableAllBreakpointsAction extends AbstractRemoveAllActionDelegate
 
 	@Override
 	public void run(IAction action) {
-		final IBreakpointManager breakpointManager = DebugPlugin.getDefault().getBreakpointManager();
-		final IBreakpoint[] breakpoints = breakpointManager.getBreakpoints();
-		if (breakpoints.length < 1) {
-			return;
-		}
-		IWorkbenchWindow window = DebugUIPlugin.getActiveWorkbenchWindow();
-		if (window == null) {
-			return;
-		}
-		IPreferenceStore store = DebugUIPlugin.getDefault().getPreferenceStore();
-		boolean prompt = store.getBoolean(IDebugPreferenceConstants.PREF_PROMPT_DISABLE_ALL_BREAKPOINTS);
-		boolean proceed = true;
-		if (prompt) {
-			MessageDialogWithToggle mdwt = MessageDialogWithToggle.openYesNoQuestion(window.getShell(),
-					ActionMessages.DisableAllBreakPointsAction_0, ActionMessages.DisableAllBreakPointsAction_1,
-					ActionMessages.DisableAllBreakPointsAction_2, !prompt, null, null);
-			if (mdwt.getReturnCode() != IDialogConstants.YES_ID) {
-				proceed = false;
-			} else {
-				store.setValue(IDebugPreferenceConstants.PREF_PROMPT_DISABLE_ALL_BREAKPOINTS,
-						!mdwt.getToggleState());
-			}
-		}
-		if (proceed) {
-			new Job(ActionMessages.DisableAllBreakPointsAction_1) {
-				@Override
-				protected IStatus run(IProgressMonitor monitor) {
-					try {
-						for (IBreakpoint breakpoint : breakpoints) {
-							breakpoint.setEnabled(false);
-						}
-					} catch (CoreException e) {
-						DebugUIPlugin.log(e);
-						return Status.CANCEL_STATUS;
-					}
-					return Status.OK_STATUS;
-				}
-			}.schedule();
-		}
+		updateBreakpoints(false);
 	}
 }
